+++ src/clients/graphql/test/graphql_client.test.ts
@@ -4,6 +4,7 @@ import {assertHttpRequest} from '../../http_client/test/test_helper';
 import {GraphqlClient} from '../graphql_client';
 import {Context} from '../../../context';
 import * as ShopifyErrors from '../../../error';
+import {ApiClientType} from '../../http_client/types';
 
 const DOMAIN = 'shop.myshopify.com';
 const QUERY = `
@@ -23,8 +24,13 @@ const successResponse = {
 };
 
 describe('GraphQL client', () => {
-  it('can return response', async () => {
-    const client: GraphqlClient = new GraphqlClient(DOMAIN, 'bork');
+  it('can return response from Admin API by default', async () => {
+    const client: GraphqlClient = new GraphqlClient({
+      domain: DOMAIN,
+      accessToken: 'bork',
+    });
+    expect(client.apiType).toEqual(ApiClientType.Admin);
+
     fetchMock.mockResponseOnce(JSON.stringify(successResponse));
 
     await expect(client.query({data: QUERY})).resolves.toEqual(buildExpectedResponse(successResponse));
@@ -36,8 +42,40 @@ describe('GraphQL client', () => {
     });
   });
 
+  it('can return response from Storefront API', async () => {
+    const client: GraphqlClient = new GraphqlClient({
+      apiType: ApiClientType.Storefront,
+      domain: DOMAIN,
+      accessToken: 'bork',
+    });
+    expect(client.apiType).toEqual(ApiClientType.Storefront);
+
+    fetchMock.mockResponseOnce(JSON.stringify(successResponse));
+
+    await expect(client.query({data: QUERY})).resolves.toEqual(buildExpectedResponse(successResponse));
+    assertHttpRequest({
+      method: 'POST',
+      domain: DOMAIN,
+      path: '/api/unstable/graphql.json',
+      data: QUERY,
+    });
+  });
+
+  it('fails with invalid API type', async () => {
+    const client = new GraphqlClient({
+      apiType: 'Invalid type!' as ApiClientType,
+      domain: DOMAIN,
+      accessToken: 'dummy-token',
+    });
+
+    await expect(client.query({data: QUERY})).rejects.toThrow(ShopifyErrors.ShopifyError);
+  });
+
   it('merges custom headers with default', async () => {
-    const client: GraphqlClient = new GraphqlClient(DOMAIN, 'bork');
+    const client: GraphqlClient = new GraphqlClient({
+      domain: DOMAIN,
+      accessToken: 'bork',
+    });
     const customHeader: Record<string, string> = {
       'X-Glib-Glob': 'goobers',
     };
@@ -62,13 +100,14 @@ describe('GraphQL client', () => {
     Context.IS_PRIVATE_APP = true;
     Context.initialize(Context);
 
-    const client: GraphqlClient = new GraphqlClient(DOMAIN);
+    const client: GraphqlClient = new GraphqlClient({domain: DOMAIN});
     fetchMock.mockResponseOnce(JSON.stringify(successResponse));
+
+    await expect(client.query({data: QUERY})).resolves.toEqual(buildExpectedResponse(successResponse));
 
     const customHeaders: Record<string, string> = {};
     customHeaders[ShopifyHeader.AccessToken] = 'test_secret_key';
 
-    await expect(client.query({data: QUERY})).resolves.toEqual(buildExpectedResponse(successResponse));
     assertHttpRequest({
       method: 'POST',
       domain: DOMAIN,
@@ -78,12 +117,28 @@ describe('GraphQL client', () => {
     });
   });
 
+  it('fails for private apps without a token', async () => {
+    Context.IS_PRIVATE_APP = true;
+    Context.PRIVATE_APP_STOREFRONT_ACCESS_TOKEN = undefined;
+    Context.initialize(Context);
+
+    const client: GraphqlClient = new GraphqlClient({
+      apiType: ApiClientType.Storefront,
+      domain: DOMAIN,
+    });
+
+    await expect(client.query({data: QUERY})).rejects.toThrow(ShopifyErrors.ShopifyError);
+  });
+
   it('fails to instantiate without access token', () => {
-    expect(() => new GraphqlClient(DOMAIN)).toThrow(ShopifyErrors.MissingRequiredArgument);
+    expect(() => new GraphqlClient({domain: DOMAIN})).toThrow(ShopifyErrors.MissingRequiredArgument);
   });
 
   it('can handle queries with variables', async () => {
-    const client: GraphqlClient = new GraphqlClient(DOMAIN, 'bork');
+    const client: GraphqlClient = new GraphqlClient({
+      domain: DOMAIN,
+      accessToken: 'bork',
+    });
     const queryWithVariables = {
       query: `query FirstTwo($first: Int) {
         products(first: $first) {