+++ src/clients/rest/rest_client.ts
@@ -3,27 +3,39 @@ import querystring from 'querystring';
 import {Context} from '../../context';
 import {ShopifyHeader} from '../../base_types';
 import {HttpClient} from '../http_client/http_client';
-import {RequestParams, GetRequestParams} from '../http_client/types';
+import {ApiClientParams, ApiClientType, RequestParams, GetRequestParams} from '../http_client/types';
 import * as ShopifyErrors from '../../error';
 
 import {RestRequestReturn, PageInfo} from './types';
 
+interface AccessTokenHeader {
+  header: string;
+  value: string;
+}
+
 class RestClient extends HttpClient {
   private static LINK_HEADER_REGEXP = /<([^<]+)>; rel="([^"]+)"/;
 
-  public constructor(domain: string, readonly accessToken?: string) {
-    super(domain);
+  readonly apiType: ApiClientType;
+  readonly accessToken?: string;
 
-    if (!Context.IS_PRIVATE_APP && !accessToken) {
+  public constructor(params: ApiClientParams) {
+    super(params.domain);
+
+    this.apiType = params.apiType || ApiClientType.Admin;
+    this.accessToken = params.accessToken;
+
+    if (!Context.IS_PRIVATE_APP && !this.accessToken) {
       throw new ShopifyErrors.MissingRequiredArgument('Missing access token when creating REST client');
     }
   }
 
   protected async request(params: RequestParams): Promise<RestRequestReturn> {
-    params.extraHeaders = {...params.extraHeaders};
-
-    params.extraHeaders[ShopifyHeader.AccessToken] = Context.IS_PRIVATE_APP
-      ? Context.API_SECRET_KEY : this.accessToken as string;
+    const accessTokenHeader: AccessTokenHeader = this.getAccessTokenHeaderForApiType();
+    params.extraHeaders = {
+      [accessTokenHeader.header]: accessTokenHeader.value,
+      ...params.extraHeaders,
+    };
 
     params.path = this.getRestPath(params.path);
 
@@ -75,16 +87,51 @@ class RestClient extends HttpClient {
   }
 
   private getRestPath(path: string): string {
-    return `/admin/api/${Context.API_VERSION}/${path}.json`;
+    return `${this.getBasePathForApiType()}/${Context.API_VERSION}/${path}.json`;
   }
 
   private buildRequestParams(newPageUrl: string): GetRequestParams {
+    const pattern = `^${this.getBasePathForApiType()}/[^/]+/(.*).json$`;
+
     const url = new URL(newPageUrl);
-    const path = url.pathname.replace(/^\/admin\/api\/[^/]+\/(.*)\.json$/, '$1');
+    const path = url.pathname.replace(new RegExp(pattern), '$1');
     const query = querystring.decode(url.search.replace(/^\?(.*)/, '$1')) as Record<string, string | number>;
     return {
       path,
       query,
+    };
+  }
+
+  private getBasePathForApiType(): string {
+    switch (this.apiType) {
+      case ApiClientType.Admin:
+        return '/admin/api';
+      default:
+        throw new ShopifyErrors.ShopifyError(`Unsupported REST API client type '${this.apiType}'`);
+    }
+  }
+
+  private getAccessTokenHeaderForApiType(): AccessTokenHeader {
+    let header: string;
+    let value: string | undefined;
+    switch (this.apiType) {
+      case ApiClientType.Admin:
+        header = ShopifyHeader.AccessToken;
+        value = Context.IS_PRIVATE_APP ? Context.API_SECRET_KEY : this.accessToken;
+        break;
+      default:
+        throw new ShopifyErrors.ShopifyError(`Unsupported REST API client type '${this.apiType}'`);
+    }
+
+    if (!value) {
+      throw new ShopifyErrors.ShopifyError(
+        `Could not determine the access token header for API client type '${this.apiType}'`,
+      );
+    }
+
+    return {
+      header,
+      value,
     };
   }
 }