+++ src/clients/rest/test/rest_client.test.ts
@@ -2,7 +2,7 @@ import '../../../test/test_helper';
 import querystring from 'querystring';
 
 import {ShopifyHeader} from '../../../base_types';
-import {DataType, GetRequestParams} from '../../http_client/types';
+import {ApiClientType, DataType, GetRequestParams} from '../../http_client/types';
 import {assertHttpRequest} from '../../http_client/test/test_helper';
 import {RestClient} from '../rest_client';
 import {RestRequestReturn, PageInfo} from '../types';
@@ -20,8 +20,12 @@ const successResponse = {
 };
 
 describe('REST client', () => {
-  it('can make GET request', async () => {
-    const client = new RestClient(domain, 'dummy-token');
+  it('can make GET request to Admin API by default', async () => {
+    const client = new RestClient({
+      domain,
+      accessToken: 'dummy-token',
+    });
+    expect(client.apiType).toEqual(ApiClientType.Admin);
 
     fetchMock.mockResponseOnce(JSON.stringify(successResponse));
 
@@ -29,8 +33,21 @@ describe('REST client', () => {
     assertHttpRequest({method: 'GET', domain, path: '/admin/api/unstable/products.json'});
   });
 
+  it('fails with invalid API type', async () => {
+    const client = new RestClient({
+      apiType: ApiClientType.Storefront,
+      domain,
+      accessToken: 'dummy-token',
+    });
+
+    await expect(client.get({path: 'products'})).rejects.toThrow(ShopifyErrors.ShopifyError);
+  });
+
   it('can make POST request with JSON data', async () => {
-    const client = new RestClient(domain, 'dummy-token');
+    const client = new RestClient({
+      domain,
+      accessToken: 'dummy-token',
+    });
 
     fetchMock.mockResponseOnce(JSON.stringify(successResponse));
 
@@ -52,7 +69,10 @@ describe('REST client', () => {
   });
 
   it('can make POST request with form data', async () => {
-    const client = new RestClient(domain, 'dummy-token');
+    const client = new RestClient({
+      domain,
+      accessToken: 'dummy-token',
+    });
 
     fetchMock.mockResponseOnce(JSON.stringify(successResponse));
 
@@ -74,7 +94,10 @@ describe('REST client', () => {
   });
 
   it('can make PUT request with JSON data', async () => {
-    const client = new RestClient(domain, 'dummy-token');
+    const client = new RestClient({
+      domain,
+      accessToken: 'dummy-token',
+    });
 
     fetchMock.mockResponseOnce(JSON.stringify(successResponse));
 
@@ -96,7 +119,10 @@ describe('REST client', () => {
   });
 
   it('can make DELETE request', async () => {
-    const client = new RestClient(domain, 'dummy-token');
+    const client = new RestClient({
+      domain,
+      accessToken: 'dummy-token',
+    });
 
     fetchMock.mockResponseOnce(JSON.stringify(successResponse));
 
@@ -107,7 +133,10 @@ describe('REST client', () => {
   });
 
   it('merges custom headers with the default ones', async () => {
-    const client = new RestClient(domain, 'dummy-token');
+    const client = new RestClient({
+      domain,
+      accessToken: 'dummy-token',
+    });
 
     const customHeaders: Record<string, string> = {
       'X-Not-A-Real-Header': 'some_value',
@@ -125,8 +154,15 @@ describe('REST client', () => {
 
   it('includes pageInfo of type PageInfo in the returned object for calls with next or previous pages', async () => {
     const params = getDefaultPageInfo();
-    const client = new RestClient(domain, 'dummy-token');
-    const linkHeaders = [`<${params.previousPageUrl}>; rel="previous"`, `<${params.nextPageUrl}>; rel="next"`];
+    const client = new RestClient({
+      domain,
+      accessToken: 'dummy-token',
+    });
+    const linkHeaders = [
+      `<${params.previousPageUrl}>; rel="previous"`,
+      `<${params.nextPageUrl}>; rel="next"`,
+      'This invalid info header will be ignored',
+    ];
 
     fetchMock.mockResponses([JSON.stringify(successResponse), {headers: {link: linkHeaders.join(', ')}}]);
 
@@ -138,7 +174,10 @@ describe('REST client', () => {
 
   it('is able to make subsequent get requests to either pageInfo.nextPage or pageInfo.prevPage', async () => {
     const params = getDefaultPageInfo();
-    const client = new RestClient(domain, 'dummy-token');
+    const client = new RestClient({
+      domain,
+      accessToken: 'dummy-token',
+    });
     const linkHeaders = [`<${params.previousPageUrl}>; rel="previous"`, `<${params.nextPageUrl}>; rel="next"`];
 
     fetchMock.mockResponses(
@@ -161,7 +200,10 @@ describe('REST client', () => {
 
   it('can request next pages until they run out', async () => {
     const params = getDefaultPageInfo();
-    const client = new RestClient(domain, 'dummy-token');
+    const client = new RestClient({
+      domain,
+      accessToken: 'dummy-token',
+    });
     const linkHeaders = [`<${params.previousPageUrl}>; rel="previous"`, `<${params.nextPageUrl}>; rel="next"`];
 
     fetchMock.mockResponses(
@@ -181,7 +223,10 @@ describe('REST client', () => {
 
   it('can request previous pages until they run out', async () => {
     const params = getDefaultPageInfo();
-    const client = new RestClient(domain, 'dummy-token');
+    const client = new RestClient({
+      domain,
+      accessToken: 'dummy-token',
+    });
     const linkHeaders = [`<${params.previousPageUrl}>; rel="previous"`, `<${params.nextPageUrl}>; rel="next"`];
 
     fetchMock.mockResponses(
@@ -203,14 +248,15 @@ describe('REST client', () => {
     Context.IS_PRIVATE_APP = true;
     Context.initialize(Context);
 
-    const client = new RestClient(domain);
+    const client = new RestClient({domain});
 
     fetchMock.mockResponseOnce(JSON.stringify(successResponse));
+
+    await expect(client.get({path: 'products'})).resolves.toEqual(buildExpectedResponse(successResponse));
 
     const customHeaders: Record<string, string> = {};
     customHeaders[ShopifyHeader.AccessToken] = 'test_secret_key';
 
-    await expect(client.get({path: 'products'})).resolves.toEqual(buildExpectedResponse(successResponse));
     assertHttpRequest({
       method: 'GET',
       domain,
@@ -219,8 +265,19 @@ describe('REST client', () => {
     });
   });
 
+  it('fails for private apps without a secret key', async () => {
+    Context.IS_PRIVATE_APP = true;
+    Context.API_SECRET_KEY = '';
+
+    const client = new RestClient({domain});
+
+    fetchMock.mockResponseOnce(JSON.stringify(successResponse));
+
+    await expect(client.get({path: 'products'})).rejects.toThrow(ShopifyErrors.ShopifyError);
+  });
+
   it('fails to instantiate without access token', () => {
-    expect(() => new RestClient(domain)).toThrow(ShopifyErrors.MissingRequiredArgument);
+    expect(() => new RestClient({domain})).toThrow(ShopifyErrors.MissingRequiredArgument);
   });
 });